Move PublishedGradleVersionWrapper to util.gradle package
